+++ non_catalog_apps/iconedit/panels/canvas.h
@@ -1,6 +1,9 @@
 #pragma once
 
 #include <strings.h> // for size_t
+#include "../icon.h"
 
-void canvas_alloc_canvas(size_t width, size_t height);
+void canvas_initialize(IEIcon* icon, size_t scale);
 void canvas_free_canvas();
+
+void canvas_set_scale(size_t scale_setting);